requirements.txt
asgiref==3.2.3
astroid==2.3.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
hipo-drf-exceptions==0.1.3
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
reportlab==3.5.42
six==1.14.0
sqlparse==0.3.0
typed-ast==1.4.1
wrapt==1.11.2
